Remove Encrypted Jobs Files (#15116)
* replace dependents encrypted jobs with regular jobs except code

* lint fix

* Fix some renames

* Other messed up renames

* Fix weird rebase issue for other flipper logic

* Remove encrypted job files and move code into normal named files

* One missed class name in spec
